pyproject.toml
[project]
name = "demo-repository"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = []

[dependency-groups]
dev = [
    "pyyaml>=6.0.3",
    "ruff>=0.15.12",
]
